+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Scheduling notifications if sending is between 11pm and 6am
 * Notifications for errored connector are now sent at 6am instead of 8am
+* Show message when request failed with refresh button
 
 ## 🐛 Bug Fixes
 
@@ -12,7 +13,7 @@
 
 ## 🔧 Tech
 
-*
+* Add RequestStateProvider to the application
 
 # 1.38.0